requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
